feat: 챌린지 방장 넘기기 기능 추가

+++ src/main/java/goojeans/harulog/challenge/service/ChallengeServiceImpl.java
@@ -124,6 +124,10 @@ public class ChallengeServiceImpl implements ChallengeService {
         Challenge challenge = challengeRepository.findByChallengeId(request.getChallengeId()).orElseThrow(() -> new BusinessException(ResponseCode.CHALLENGE_NOT_FOUND));
         ChallengeUser challengeUser = challengeUserRepository.findChallengeUserByUserAndChallenge(user.getId(), challenge.getChallengeId()).orElseThrow(() -> new BusinessException(ResponseCode.CHALLENGE_NO_PERMISSION));
 
+        if (challengeUser.getRole().equals(ChallengeRole.LEADER)) {
+            throw new BusinessException(ResponseCode.CHALLENGE_LEADER_CANNOT_LEAVE);
+        }
+
         challenge.removeChallengeUser(challengeUser);
         challengeUserRepository.delete(challengeUser);
 
@@ -275,6 +279,25 @@ public class ChallengeServiceImpl implements ChallengeService {
 
             challenge.removeChallengeUser(kickoutChallengeUser);
             challengeUserRepository.delete(kickoutChallengeUser);
+
+            return Response.ok();
+        } else {
+            throw new BusinessException(ResponseCode.CHALLENGE_UNAUTHORIZED_ACCESS);
+        }
+    }
+
+    @Override
+    public Response<Void> assignLeader(Long userId, Long challengeId, String nickname) {
+
+        boolean isChallengeLeader = isChallengeLeader(userId, challengeId);
+
+        if (isChallengeLeader) {
+            Users user = userRepository.findByNickname(nickname).orElseThrow(() -> new BusinessException(ResponseCode.USER_NOT_FOUND));
+            ChallengeUser challengeUser = challengeUserRepository.findChallengeUserByUserAndChallenge(user.getId(), challengeId).orElseThrow(() -> new BusinessException(ResponseCode.CHALLENGE_NO_PERMISSION));
+            ChallengeUser challengeLeader = challengeUserRepository.findChallengeUserByUserAndChallenge(userId, challengeId).orElseThrow(() -> new BusinessException(ResponseCode.CHALLENGE_NO_PERMISSION));;
+
+            challengeLeader.updateRole();
+            challengeUser.updateRole();
 
             return Response.ok();
         } else {
